drop(integration): Always fetch max number of pages
It is very easy to write code and not pass the parameter, thus, we won't get the whole set of repositories.

This will cause the calls to be slower for customers with thousands of repos, however, it will be more accurate.

We still have `self.page_number_limit` to stop the requests after 50 page fetches (5,000 repositories).